添加部门删除修改功能

+++ src/main/java/com/xinguan/usermanage/service/impl/DepartmentServiceImpl.java
@@ -6,6 +6,9 @@ import com.xinguan.usermanage.model.Employee;
 import com.xinguan.usermanage.model.Role;
 import com.xinguan.usermanage.service.DepartmentService;
 import org.apache.commons.lang3.StringUtils;
+import org.assertj.core.util.Sets;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.data.domain.Example;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
@@ -17,7 +20,10 @@ import org.springframework.util.Assert;
 
 import javax.persistence.criteria.Predicate;
 import java.util.Date;
+import java.util.List;
 import java.util.Optional;
+import java.util.Set;
+import java.util.stream.Collectors;
 
 /**
  * @author zhangzhan
@@ -25,6 +31,7 @@ import java.util.Optional;
 @Service
 public class DepartmentServiceImpl extends BaseService<Department> implements DepartmentService {
 
+    private static final Logger LOGGER = LoggerFactory.getLogger(DepartmentServiceImpl.class);
 
     @Override
     public Page<Department> listDepartmentByPage(int pageSize, int pageNo, String departmentName, Employee employee) {
@@ -68,7 +75,19 @@ public class DepartmentServiceImpl extends BaseService<Department> implements De
 
     @Override
     @Transactional
-    public Department saveOrUpdate(Department department) {
+    public Department saveOrUpdate(Department department, Long superiorDepartmentId) {
+
+        //上级部门
+        if (superiorDepartmentId != null) {
+            Department superDepartment = departmentRepository.getOne(superiorDepartmentId);
+
+            if (superDepartment != null) {
+                department.setSuperiorDepartment(superDepartment);
+            } else {
+                LOGGER.warn("not found department,superiorDepartmentId:[" + superiorDepartmentId + "]");
+            }
+        }
+
         Example<Department> departmentExample = getSimpleExample(department);
         if (departmentRepository.exists(departmentExample)) {
             department.setModificationDate(new Date());
@@ -83,9 +102,27 @@ public class DepartmentServiceImpl extends BaseService<Department> implements De
         return departmentRepository.getOne(id);
     }
 
+
     @Transactional
     @Override
-    public void removeDepartment(Long id) {
-        departmentRepository.deleteById(id);
+    public void removeDepartmentBatch(String ids) {
+        if (StringUtils.isNotBlank(ids)) {
+            Set<Long> longIdList = Sets.newTreeSet(ids.split(",")).stream().map(Long::parseLong).collect(Collectors.toSet());
+            longIdList.forEach(id -> {
+                Department department = departmentRepository.getOne(id);
+                List<Employee> employees = employeeRepository.findAllByDepartment(department);
+                employees.forEach(employee -> {
+                    employee.setDepartment(null);
+                    employeeRepository.saveAndFlush(employee);
+                });
+                department.setSuperiorDepartment(null);
+                departmentRepository.saveAndFlush(department);
+                departmentRepository.deleteById(id);
+            });
+        }
+
     }
+
+
+
 }
